requirements.txt
aiohttp==3.12.13
beautifulsoup4==4.13.4
fastapi==0.115.13
jose==1.0.0
livekit==1.0.9
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation==0.2.4
livekit_api==1.0.2
livekit_protocol==1.0.3
numpy==2.3.0
openai==1.88.0
passlib
psycopg2
pydantic[email]
pydantic
PyJWT
python-docx
python-dotenv
python-multipart
python_jose
pytz
PyMuPDF
requests
SQLAlchemy
starlette
uvicorn
bcrypt
livekit-agents[elevenlabs]~=1.0
